Ported Data::Random to use Time::Piece
Time::Piece is a core module since Perl 5.10.
Date::Calc is diffucult to install and not so nice.

And apart from that since 1st of January 2015 the module started to
fail its unit tests and this has not been fixed yet.

https://rt.cpan.org/Public/Bug/Display.html?id=101232

I modified just the module in this commit. The test files also use
Date::Calc but they're in a separate commit so you can easily check
out just the module changes and see the 'old' tests pass with the
new code.