Port development to Dist::Zilla
This patch moves development to Dist::Zilla, in particular to solve
issues with the removal of . from INC, but more generally to make
several tasks related to package authoring easier.

The plugin list used has been kept to a minimum, while still
trying to reproduce some of the functionality that was there in the
previous configuration, and make use of some of the more useful
features of MetaCPAN, like the contributor list.

To this end, a .mailmap file is also added, to remove non-ASCII
characters from contributor's names. And yes, this is 2018.